unify RPS so easier to run directly
